+++ app/src/helpers/getPrisma.ts
@@ -2,7 +2,7 @@ import { PrismaClient } from '@prisma/client'
 
 let prisma: PrismaClient
 
-export default function getPrisma() {
+export default function getPrisma(): PrismaClient {
   if (!prisma) {
     prisma = new PrismaClient()
   }
